(#42354153) Move HTTP/HTTPS fw rules to Nginx
Move firewall rules for HTTP and HTTPS to the Nginx module.

DANGER DANGER:

As best as I can tell, Nginx is the only thing that binds to HTTP and HTTPS
directly. Everything else is presented through Nginx vhosts and proxy_pass.
However somebody with more knowledge of the stack should double check that
fact before merging.