Remove healthcheck code from nginx configs
Healthchecks are no longer a critical part of our load balancing. This commit
removes any special configuration of healthcheck ports/servers/upstreams from
our nginx configs.

`govuk::app`s still define a `$health_check_path`, which is used to export a
Nagios check which ensures that the app is returning a sensible HTTP status
code at the specified path, but as this path should be served alongside the
rest of the application, this no longer requires custom nginx setup.

I've updated the documentation of the `health_check_path` param in
`govuk::app` to reflect this new usage.