(#42354153) Re-allow multiple es::node invocations
The move of the firewall rule from govuk::firewall to es::node::firewall
removed the ability to for a single Puppet node to invoke multiple instances
of es::node due to the use of a param class and resource naming.

Re-allow this, by converting the class to a define to match the other
es::node::* resources and name the firewall resource after the port that has
been passed in. It will rightly disallow multiple ES instances from using
the same http_port on the same machine.