(#43551733) Use JS for monitoring splash links
Use client-side javascript on the monitoring splash page to generate
domain-relative links instead of rendering them server-side with an
environment specific template. This makes "monitoring.staging" go to the
right place, instead of linking out to production. It also frees us from the
use of `app_domain` in this class.

No bones about it, this is still pretty horrible, but in the grand scheme of
things it is a little bit better and agnostic of physical environments.