Passive check for puppet runs on all machines
This commit takes the code from 0c0f0d3e714f0, which adds a passive
check for puppet errors on the monitoring box only, and generalizes it
to work with all machines by sending checks via NSCA.

Interestingly, this only works for 'puppet agent' and not for 'puppet
apply' which we use in our vagrant environments, because we have
specified the postrun_command in the [agent] rather than the [main]
section of the puppet.conf. Not sure what the consequences of changing
this are, so I'm leaving it for the moment.