Remove all traces of Jetty
Jetty was used to run the old Scala based router on the "cache" nodes. This
was later replaced by Varnish. The last includes for Jetty were cleared up
in 11b94be and 448daa6.

The only Java apps now in use are Play2 based which use the bundled Netty
server and don't require as much fiddling. So removing this module and
associated sudo/log files for good housekeeping.

I expect that @samjsharpe will be pleased.