(#43475347) Framework for collectd plugins
The principle is that each module (like Varnish) will include a
`collectd::plugin::x` sub-class for it's particular collectd plugin. These
are virtual resources so that there isn't a hard dependency on collectd - we
wouldn't want to run this on the development VM for example.

The main `collectd` class then hoovers up all of these virtual resources for
machines that it is installed on.