(#44008625) Refactor $varnish_version and tests
Encapsulate the variable within `varnish::config`. It's only used by a
template contained within this class, so we don't have to look it up across
scopes. This makes it more clear where it's coming from and what it's being
used by.

It also means that the tests can (and have) be split apart for the separate
classes. Which reduces the runtime for individual contexts.

Minor style changes:

- Rename `$varnish_version` to `$vcl_version`.
- Replace case statement with selector.