Set a more sensible heap size and don't replicate
We currently use the default ES heap size of 512M in production, which is a
bit worrying, as our indices are already about 230M in size. In addition, we
have number_of_replicas set to 1, which is pointless with a single node
cluster.

This commit fixes both of these things.