(#44030613) Allow /nginx_status from 127.0.0.0/24
Nginx stub_status is bound and restricted to loopback for security purposes.
For a little extra obscurity it uses `127.0.0.234` instead of `127.0.0.1`

This worked fine on Lucid, because it appears that when using loopback the
2.6 kernel matches the source address to the destination address. However
Precise/3.2 always defaults to a source address of `127.0.0.1` and causes
Nginx to 403 the requests.

Keep the obscurity of binding to the non-standard address. But relax the
restriction of source addresses to `127.0.0.0/24` in order to allow
`127.0.0.1` and `127.0.0.234`. Technically it could be `127/8` to match the
interface, but I see no reason to make it that wide.