(#39116501) Add /healthcheck res to Nginx default
Add a `/healthcheck` resource to all Nginx default vhosts which will return
a static 200 response when Nginx is up. The URI can be modified, or disabled
completely, using the param.

This will satisfy vShield Edge health checks which don't supply a `Host`
header. It's not possible for VSE to test Varnish behind Nginx, because it
won't follow the 30x HTTP->HTTPS and it only does SSL handshakes.