Refactor elasticsearch module: ES node per server
The elasticsearch module previously allowed multiple instantiations of
elasticsearch on a single server. In practice, this facility was never used.

This commit removes the elasticsearch::node type, reverting to the use of the
elasticsearch class to define elasticsearch nodes, thus simplifying the
configuration of elasticsearch plugins (by, for example, making it obvious
which service needs to be restarted).

This commit does not remove the naming conventions of various directories
which depend on the $cluster_name passed to the elasticsearch class. Nor does
it rename the service. This hopefully means that it won't touch existing
elasticsearch instances, and the problem of removing/renaming these resources
is left as an exercise for the future.