(#44368133) Switch nginx.pid path on >=12.04
12.04 relocated `/var/run` to `/run`. These is a symlink in place, but Nginx
doesn't appear to respect it. At some point during config/restart this
causes Nginx to remove the PID but continue running.

The result of this is that Puppet continually tries to start the service
again and logrotate fails to tell the process to reopen log files when
they've been rorated.

Template the file and prepend `/var` on Lucid nodes. Anything later
shouldn't get the prefix.