[#44519757] Prepend PATH for govuk_puppet
The `govuk_puppet` cronjob isn't working on newly provisioned Precise
machines cache-{1,2}.router.preview and the following error is going to
root's mail spool:

    /usr/local/bin/govuk_puppet: 6: exec: agent: not found

This is because the binstub for Puppet is in `/usr/local/bin` and this isn't
on cron's default PATH of `/usr/bin:/bin`. It's not affecting old machines
because their binstub is in `/usr/bin`.

Fix this by prepending to PATH within `govuk_puppet`.