[#44000493] Add _credentials to extlookup hierachy
Add a new entry to extlookup's search hierachy. Allow credentials like
passwords and keys to be split out to a separate file. Making extdata look
like:

- production_credentials.csv
- production.csv
- common.csv

The rationale for this is that tracking changes between the different files
is hard. Even moreso when they come to be split out to separate repos. The
biggest risk is that feature flags are enabled for one environment and
forgotten about for another.

Moving the credentials (which should be present in most environments) should
make it easier to track the feature flags (many of which will only be
enabled in one or two environments).

This commit is backwards compatible. The extdata repos and Capistrano jobs
can be restructed afterwards.