Restrict dependencies on dev VM
Group the gems added in 66f6739 for testing the mirrorer so that the dev VM
can exclude them. This resolves an issue whereby it was failing to install
nokogiri because of library dependencies on libxml2.

I am inclined to think that the longer term solution to this is to create a
Vagrant middleware that updates the repos and librarian-puppet outside of
the VM. Rather than sharing state with directories on the host and guest.