[#44000493] Make puppet environment sticky
Populate the environment in puppet.conf from Puppet's internal
`$::environment` variable. This will default to "production" in most places.

Most of the time it won't change, however it can be updated in a single run
with `govuk_puppet --environment foo`. Will enable us to transition the
development and vagrant-govuk VMs onto a different environment and set of
extdata files.