[#44000493] Change extdata path to be relative
Use a single path for extdata that is parallel to the manifests and modules
directories in this repo. This matches the path created by Capistrano in our
real deployments. Sensitive CSVs from the deployment repo will get copied on
top of this.

vagrant-govuk and development will need to set the appropriate
`--environment` argument to pick up the correct CSV files. vagrant-govuk
will also need fresh CSVs created and the extdata mount correcting.