[#44000493] Copy wildcard_alphagov dev->vagrant
The wildcard_alphagov cert has a CN of `*.dev.alphagov.co.uk` and is signed
by Comodo. It is currently used by the development VM and is deemed
relatively low risk. Re-use this for vagrant-govuk so that we don't have to
deal with self-signed cert trust issues.

ACKed by @samjsharpe