[#45048111] Import logster v0.0.1
Importing the codebase of logster from it's v0.0.1 tag on github:etsy/logster

This is presently unreferenced by the Puppet manifests. It will be used as a
base for any in-house customisations that we need to port over, then the
manifests can be switched over and the new version used.