[#45048111] Nasty hack for logster imports
Logster now expects to be installed as a Python module and simply qualifies
it's includes with `logster.`. However they recommend you install it with
`python setup.py install` and haven't published to PyPI *trollface*

Workaround this in the least horrible manner by moving everything up one
directory (so we don't search all of /usr/share) and appending the path to
`sys.path` as before.