Remove collectd prefix from Graphite metric keys
I originally thought it would be a good idea to prefix all Graphite metrics
that were originating from collectd with `collectd.` to distinguish them
from the `ganglia.` and `statsd.` namespaces.

However it's occurred to me that a handful of collectors that currently
write via Ganglia will soon be writing to Graphite directly and don't belong
under collectd's namespace. Introducing a new top-level key for each of
these is going to be messy - not least because it makes hard to browse all
metrics available for a single host.

Removing this key so that everything gets rooted to the node's FQDN. Any
existing whisper files created by collectd will need to be moved or deleted.
Probably the latter, since we only have a couple of day's worth of data.

Further notes:

1. Statsd and Logster should be reconfigured at the appropriate time so that
  they write directly into each FQDN's namespace. Assuming that we don't
  care where a particular key has come from ie. retention rates should be
  based on the data not the source.

1. We could do something fancy on each of the collectd clients to prefix
  with a reversed FQDN. This would create a DNS based hierachy such as:

  - production
    - frontend
      - frontend-1
      - frontend-2
    - backend
      - backend-1
      - backend-2