Reduce frequency router::errorpage execs
This curl fetches rendered error pages from the static app to Nginx's
webroot on the cache/router nodes. These are used to serve friendly and
styled user-facing errors, either intercepted or as the result of an
upstream failure.

The exec has no conditions, which causes it to be triggered many times on
every Puppet run. However these pages[0] don't actually change very often.

Reduce the exec to once every 6 hours, or if the files don't exist. This
will reduce the overhead and noise of every Puppet run. It will also make
reports and metrics more manageable should we choose to digest them in the
future.

[0] https://github.com/alphagov/static/tree/master/app/views/root