[#45048111] DRY up logster::cronjob arguments
The arguments to logster commands are spread and duplicated across quite a
few places. This is going to make it hard to introduce additional output
specific args, such as additional prefixes to the metric name.

DRY these up by moving them into separate params on the define. These all
use the same parser so we can just set a param default.

Also be opinionated about the output by once again hardcoding "ganglia". We
need two cronjobs to transition to Graphite. I'm going to repurpose the
existing cron resource to write to Graphite and create a temporary one that
writes to Ganglia.