[#45048111] Temporarily add custom arg to logster
Logster allows you to write to Ganglia and Graphite at the same time.
However the metric names will not be correct. Ganglia automatically prepends
the hostname when it is submitted locally, but for Graphite they just turn
up in the root of the tree.

To workaround this I'm going a bad thing. Forking our copy of Logster to add
a new argument, which will prepend a string only when submitting to
Graphite. This will allow the single job to submit to both data stores
during a transition period.

Once we're ready to teardown Ganglia this commit can be reverted in exchange
for putting the FQDN into `--metric-prefix`.