[#45757679] Refactor postfix (SES smarthosting)
This isolates the smarthosting functionality and will allow us to use the
same class on all hosts. As an aside, it also makes it agnostic SES,
allowing us to switch to another SMTP relay if we need to.

Variable lookups in templates from `postfix::postmapfile` are now qualified
back to `postfix::config` to prevent scope warnings. Tests should catch
these ever being changed in the class.

Should be a no-op on existing nodes.