Remove -o arg from spec parallel_tests
This sneaked in when I was testing using `--tag` to filter the spec:nodes
tests. They turned out to not be appropriate because the glob and loop in
the govuk_node spec was still being evaluated and slowed down tests.

Somehow this argument causes rspec to find more examples than we actually
have. I have no idea how. Anyway, remove it.