[#45048219] Nagios/Graphite MySQL threads/cnxs
Flip Nagios check for MySQL connections from Ganglia to Graphite.

While investigating this we discovered that the current data that we have in
Ganglia is incorrect. The `SERVER STATUS` value `connections` is a counter
of connect attempts since server start. This was being converted into a rate
per second, which never reached above 1.0

Change this to `mysql.threads-connected` from collectd's MySQL module. The
number of threads is equal to the number of connections.

The warning and critical values have been raised to account for our current
average connections of 60 on `mysql-master-1.backend` and `max_connections`
setting of 400.