[#45048219] Remove Nagios rabbitmq.messages
The data that we have in Ganglia is broken because it assumes that there
will only be a single queue. It is getting value of either the first or last
queue. Thus this alert never fires, even datainsight have some queues that
have never been consumed.

collectd is now giving us queue length metrics for every queue. Data Insight
team are interested in making use of this information. However for the
timebeing this alert is not actionable, especially by us. To be revisited in
the future.