[#45048219] Relax collectd process regex
Relax the regex used by collectd's processes plugin to match govuk::app
unicorn processes. It was believed that collectd would traverse the process
tree for child processes, however it doesn't. This might be specific to
ProcessMatch and not Process.

The regex will now match the master and worker processes. Metrics will be
summed before submission.