[#45048219] Nagios/Graphite logster HTTP codes
Flip Nagios check for HTTP status codes from Ganglia to Graphite. Logster
remains the data source, but it is now submitting to both places.

Because logster only runs every 2 minutes we have to use `keepLastValue()`
to fill in the gaps between our 10 second resolution whisper files.

This amends with the following checks:

- varnish http_5xx
- router nginx_http_5xx
- redirector nginx_http_404 and nginx_http_5xx
- puppetmaster nginx_http_5xx
- nginx::config::vhost::* http_5xx

Working on this has highlighted two issues to addess later:

1. carbon discards these datapoints when rolling up to the next retention
    period. This won't affect Nagios checks and will be addressed when we
    replace logster with logstash.
2. The key names in Graphite are pretty silly. They contain a mixture of
    duplicated and missing information. However these can't be fixed until
    we've removed Ganglia backwards compatability from logster.