[#46035347] ci.alphagov Jenkins listen on HTTPS
Make `ci.alphagov.co.uk` Jenkins master and slave instances listen on HTTPS
in addition to HTTP. So that logins and hooks can be submitted securely.

This does not force people to move over to HTTPS. See story [#46035427].

Nor have I attempted to refactor the `nginx::config::ssl` define. I'll save
that for another day. It is however documented a bit more clearly.