[#46117753] certtype for management vhosts
Add a new `certtype` for management related vhosts that aren't hosts(5)
switched between *.production and *.staging with the script
`govuk_select_organisation`.

Flip the following hostnames over to this:

- deploy.*
- nagios.*
- graphite.*
- ganglia.*
- monitoring.*

Not proud of this implementation at all. This define still needs to be
broken apart. However at least it has tests.