Speed up Graphite tests
Reduce the time that Graphite tests take by omitting sub-classes that we
aren't actually testing. We're only interested in `graphite::package`
because it varies on `$::lsbdistcodename`.

Also wrap each package in an individual `it{}` block so that `rspec -fd`
output reads nicely and individual errors are reported.

Tests down from 0.87s to 0.32s.