Use full GPG fingerprint to specify backup encryption key
It is trivially easy to generate GPG shortid collisions [1]. If we only
specify a key by its shortid, we'll download all keys from the keyserver
with that shortid.

Where it gets fun is that `gpg --encrypt -r <shortid>` will pick ONE
matching key, rather than failing because the shortid is not unique
within the keyring.

To avoid this, we should specify the whole fingerprint (or at least the
longid) when encrypting.

This won't prevent downloading keys which collide on shortid due to a
bug in gpg --recv-keys [2], but it will prevent using the wrong key when
running gpg --encrypt -r <fingerprint>.

This change is untested, due to us not running this script in any
environment except production.

I did however check the appropriate gpg commands with the colliding
keypair at A56E15A3.

[1] http://www.asheesh.org/note/debian/short-key-ids-are-bad-news.html
[2] https://bugs.g10code.com/gnupg/issue1340