[#46477001] Sanitise logster prefix
Just in case we receive something nasty. Also changed `llama_` example to
`llama_catcher` since the Graphite output does `join('.')` and it doesn't
make sense to have `llama_.http_5xx`.

Word to the wise: testing with `--output=stdout` doesn't produce the same
key names as `--output=graphite`. It's better to send the output to loopback
and watch with `nc -vlk 2003`.