[#46576471] Nagios alert for MySQL replication lag
Raise a warning and then critical when a MySQL slave falls behind it's
master by more than 5mins and 10mins respectively. We may need to tweak this
value if intensive queries cause the single-threaded slave process to block
to periods of time. But we have no prior data to go on at the moment.

This value comes from `Seconds_Behind_Master` in `SHOW SLAVE STATUS;`

- https://github.com/collectd/collectd/blob/collectd-5.1.0/src/mysql.c#L468-484

Per the MySQL documentation there is a small chance that this value might
not always be entirely accurate. However it should be fine for our needs,
where there is fast ethernet between hosts and no clients are writing
directly to the slaves:

- http://dev.mysql.com/doc/refman/5.5/en/show-slave-status.html

If replication stops altogether then a `NULL` value will be reported, which
propogate to Graphite as-is. `transformNull()` is used to convert these into
a large number (1 day). This ensures that we raise a critical rather than
"unknown" for what is an equally serious event.