[#46576471] Don't manage collectd user on slave
Now that we have some knowledge in `collectd::plugin::mysql` of whether a
node is a MySQL slave or not we can prevent the collectd MySQL account from
being created or managed on slaves.

This is necessary because the `CREATE USER` statement is propagated from the
master. If created independently on the slave then it will cause replication
to halt. This is what we saw when removing the Ganglia user (referenced in
this story).

collectd won't be able to connect to a slave, or produce any useful data,
until it has been connected to a master that has a collectd user record. It
will periodically attempt to reconnect though.