[#46886367] Generic check_graphite with args
Create a generic check_graphite handler in Nagios that can take arbitrary
arguments. This will allow us to pass `--droplast` to fix the MySQL
replication check, and similar variations in the future.

We should be able to merge `check_graphite_metric_since` into this in the
future. But not yet, because existing checks would fail until the plugin is
updated.