[#47512035] Smokey/Nagios loop-da-loop
The addition of the `-n` flag to `setlock` in 3bdf9c9 caused Smokey to exit
with a warning if an existing process was running. It became apparent from
the flood of messages to `machine.email@` that the cronjob was frequently
failing to get a lock because Smokey took >1 min (2~3) to complete.

We could increase the interval of the cronjob, but there's a likelihood that
it will need increasing again in the future. Since we effectively want to
run Smokey as often as possible, so that Nagios has the most up-to-date
results, a different approach is desireable.

Replace the cronjob with an upstart job which will re-run Smokey every time
it finishes, regardless of the exit code. Upstart's default respawn limit
will prevent it from going mad if the script is broken and exits too
quickly. We'll get a syslog entry each time it is respawned which should be
more useful than the logs from cron.

The now omitted cronjob will be purged by `resources['cron']`. We
no longer need `setlock` because upstart takes care of process management
and should never start two jobs at the same time.