[#45844199] Tests for all govuk::logstream params
We have a lot of logic within this define and it's template, so test it all.
The only thing it doesn't now test is combinations of tags and fields. Since
we're now matching the entire argument list for `govuk_logpipe` we no longer
need to check for the absence or arguments like `--json`.

Whitespace pedants may notice some issues with our current template ;)