[#47708141] Preserve FQDN in rsyslog entries
Tell rsyslog to preserve the FQDN in entries rather than using the short and
unqualified representation. This will make it easier to aggregate logs from
separate environments which have different sub-domains (like our VDCs).

There is a bug in the Ubuntu 12.04 version 5.8.6 of rsyslog that prevents it
honouring this directive on most entries; everything but the kernel and
rsyslog itself will still appear as short names.

- https://bugs.launchpad.net/ubuntu/+source/rsyslog/+bug/1022545

This might be a bit confusing, but it's a step in the right direction given:

- Most of our estate is Ubuntu 10.04 at the moment anyway.
- The fix will either be backported or we can upgrade rsyslog ourselves.
- We are adding FQDN to tagalog.