Exit govuk_logpipe if govuk_node_list fails
Bail early and with a sensible message if `govuk_logpipe` is unable to get a
list of redis nodes from `govuk_node_list`. This also applies if there are *no*
redis nodes in PuppetDB.

This was observed when running `govuk_logpipe` in a Vagrant environment
without a PuppetMaster and PuppetDB. Which results in:

    ...
    urllib2.URLError: <urlopen error [Errno 113] No route to host>
    usage: logship [-h] [-t TAGS [TAGS ...]] [-f FIELDS [FIELDS ...]] [-s SHIPPER]
                   [-j] [--source-host SOURCE_HOST] [--no-stamp] [--bulk]
                   [--bulk-index BULK_INDEX] [--bulk-type BULK_TYPE] [-k KEY]
                   [-u URLS [URLS ...]]
    logship: error: argument -u/--urls: expected at least one argument

This could also have been achieved with pipefail. However I confess to not
fully understanding the reason for prefixing the variable assignments with
`:`/`true`.

NB: This also doesn't address what the logstream upstart job does with this
output. It seems to hang indefinitely and we don't capture the output
anywhere. So we may never know if this occurs on a node.