[#48239415] Strip query params from sitemap
Strip query params from sitemap requests to ensure that it is served from
cache and only expires naturally. Generating the sitemap is currently a very
intensive process and can have a detrimental effect on other parts of the
stack. This should be reverted later this month (Apr 2013) when the platform
team have implemented changes to Rummager.

Have tested that this VCL compiles and serves the same object (`Age` header
increasing) regardless of the query params. Made it a nested if condition so
that it's not evaluated on unrelated requests and is easier to revert in
future.

/cc @mazzmosley @davidthompson @jamiecobbett