Revert "(#44377213) Make cache-* retry same on 10.04/12.04"
This reverts commit 36edebdfd24c86776b9d2ea19bd97ca946953078 and raises
`proxy_read_timeout` to 20s.

The retry bug was preserved from Lucid because we didn't understand the
impact of removing it. From the recent investigations into timeouts I think
we can now say that it is undesirable. Removing the retry and doubling the
timeout will have the same effect, whilst simplifying the request flow and
access logs.

Here's a breakdown of what previously was happened for a request that takes
15s to return from frontend/backend and isn't in cache:

1. Nginx reverse proxies a request to Varnish over loopback with a timeout
  of 10s (`proxy_read_timeout`).
2. Varnish reverse proxies the request to a VSE or Nginx at the next tier
  down with a timeout of 60s (default).
3. Nginx gives up after 10s and sends the same request to Varnish again.
4. Varnish's collapsed forwarding recognises that it's the same as the
  currently in-flight request and doesn't generate a new one.
5. Varnish receives a response to the first request and replies to the
  second Nginx request within 20s (2 * 10s).

Technically the new 20s timeout may be a few milliseconds short of the
previous behaviour, if accounting for a bit of processing between tries. It
just so happens that 20s is what we notionally set (from Nick's diagrams) as
the desired timeout to work down from. This will still leave some requests
hanging further down the stack - particularly sitemap generation, which has
been seen to take more than 20s.