[#47804651] Nagios server doesn't need exportd res
The class `monitoring::checks` is only included by the Nagios server. It
doesn't need to use exported resources to create resources for itself. Nor
does it need to pass `host_name` which defaults to `$::fqdn`.

Not only will this be less intensive to create resources, faster for them to
appear on the node, and simpler. It will also mean that they can be
conditional without having to purge from PuppetDB.