Insert carbon-aggregator in front of carbon-cache
carbon-aggregator allows us to sum or average metrics as they enter graphite. In
order to be able to reliably and conveniently track deploys, I'd like a metric
of the form

    stats.govuk.app.whitehall.all.deploys

which is the sum of

    stats.govuk.app.release.{backend-1,backend-2,backend-3}.deploys.whitehall

or similar.

With a statsd instance on each backend server, there's no nice way to let the
release application do the stats insertion without either a) three statsds
fighting over the value of the metric or b) forcing anyone who wants to plot
deploys to aggregate over three different release applications.

carbon-aggregator allows us to solve this problem by inserting a line of the
form

    stats.govuk.app.<app>.all.deploys (5) = sum stats.govuk.app.release.*.deploys.<app>

into aggregation-rules.conf.