Make puppet passive check work when run as puppet postrun command
The puppet passive check is currently broken when run as a
postrun_command. If you have a look in nagios, you can see that every
single puppet check has the status "OK: Puppet is running". This is
because (as a hangover from the active check days) it checks to see if
puppet is currently running by checking for the puppetdlock file. Since
puppet is already running, it doesn't bother checking the puppet result
yaml files. Whoops.

Since we now run as a postrun_command, when the nagios check gets run
the yaml summary files should already be in place. There is therefore no
need to check for puppetdlock. The fix is to remove the check.

No pivotal ticket as it's a literal 5 minute fix. Sorry @carlmassa

Found in coordination @danabel.