Set tempdir for clamav-daemon (fixes broken pipe)
Fixes an issue affecting Licesning after the recent upgrade of ClamAV in
PR #539. Licensing submits file contents to `clamd` over loopback on
TCP/3310 and had been logging "broken pipe" errors.

This can be reproduced with `clamdscan --stream`:

    dcarley@preview-asset-master:~$ echo foo > foo
    dcarley@preview-asset-master:~$ clamdscan --stream foo
    ERROR: Can't send to clamd: Broken pipe
    /home/dcarley/foo: no reply from clamd
    ...

Restarting `clamd` with `sudo service clamav-daemon restart` resolved this.
At first I thought this was the result of some race condition to do with the
upgrade or the way that we trigger `Exec[freshclam --quiet]` before starting
the daemon. But I couldn't reproduce this on a Vagrant VM.

Attaching `strace` to the `clamd` process whilst calling `clamdscan` on a
affected machine reveals this:

    dcarley@preview-asset-master:~$ sudo strace -vfp $(pgrep clamd)
    ...
    [pid  5279] open("/tmp/user/0/clamav-e41b103da0dd009d881b8f4777da7902", O_RDWR|O_CREAT|O_EXCL|O_TRUNC, 0700) = -1 EACCES (Permission denied)

When `clamd` receives a stream, instead of being passed a file handle, it
buffers the contents to a temporary file. However it is trying to write to a
directory that belongs only to root and `clamd` is running as the user
`clamav`.

It turns out those `/tmp/user/${UID}` directories are created by
`pam_tmpdir` for new login sessions and two environment variables are set
pointing to that directory, `$TMP` and `$TMPDIR`. That same process is
triggered each time cron run Puppet as root. The variables are passed over
to `clamd` when Puppet calls `/etc/init.d/clamav-daemon start`.

It didn't affect the Vagrant VM because Puppet is called through `sudo`
which defaults to `env_reset` and prevents any environment variables from
being passed on. Likewise my call to `sudo service` doesn't pass the env
vars, because of sudo, and `/usr/sbin/service` does `exec env -i` to
preserve a clean environment.

This could have been fixed by making Puppet use `service` instead of calling
the init script directly. But that isn't trivial with the service providers
on Debian. The most foolproof approach is to set `TemporaryDirectory` and
tell it to ignore `TMP/TMPDIR`. This config change will trigger a restart of
the daemon to fix any existing processes.

/cc @pramodbansal @davidillsley

It's not clear whether this will fix some similar problems that the
mainstream team saw when uploading assets through Travel Advice. Technically
these shouldn't have been streaming and thus not affected by the same issue.
We should test again when this has hit preview.

/cc @alextomlins @kushalpisavadia

Whitehall should remain unaffected.

/cc @bradleywright