[#48725409] apparmor::profile local_only support
Switch to a branch of our fork for the puppet-apparmor module which adds a
new `local_only` param. This allows us to manage the local profile while
delegating the contents of the parent profile to the Ubuntu package.

alphagov/puppet-apparmor@fd0106ccd688a8af4b62e57d420a26e80b484731

To be submitted upstream..