[#48566865] Remove govuk::user has_deploy param
The default value for this param was changed from `true` to `false` in
460892f to ensure that all deployments are initiated through Jenkins. At
that point all user's existing keys would have been purged from the deploy
user's `authorized_keys`. I've verified as such on frontend-1 in production.

With no `govuk::user` resources passing in a `true` value there is no longer
a need to have this param. This will make the process of managing SSH
pubkeys as `File[]` resources much easier.