[#48566865] govuk::user authorized_keys as files
Manage `authorized_keys` for `govuk::user` as file resources.

The define now takes SSH public keys as a string, or array of strings (just
for @alextomlins). The `.ssh` directory is always created while the user is
marked as present. Marking a user as absent will delete the keys but not the
directory (same behaviour as homedirs). Includes tests for some of that
weird logic.

Like the previous commit, this will enable us to reliably purge old or
unmanaged keys, which the `ssh_authorized_key{}` type doesn't allow. There
is probably a reasonable client-side speed increase now that each file only
needs to be checksummed instead of parsed.

The only disadvantage is that the contents of the keys will no longer appear
in Puppet's log output. Only the checksum of the old and new key file.
However you can still review the clientbucket to see the content changes.

NB: This commit cannot be easily reverted. We've not faking a comment on
every ssh_key entry, which is what `ssh_authorized_key` uses to identify
individual resources. This demonstrates the very problem we're trying to
solve.