[#48566865] mirror and backup SSH priv keys
Amend `mirror` and `backup` modules so that they depend on the `.ssh/`
directory that's now provided by the `govuk::user` define.

An error was picked up for the `mirror` module by the existing spec test.
Catching the `backup` module required the exhaustive `spec:nodes` test. Each
now explicitly test for the dir, rightly or wrongly, to ensure that the file
has something to autorequire upon.