Set connect timeout on router::errorpage curls
Bail if `curl` can't get a TCP `connect()` within one second. This will only
kick in if the static host is down or there are severe network connectivity
issues. It will prevent Puppet from hanging for 7 * 60s in such
circumstances, which is quite common in Vagrant if you haven't brought up a
frontend-lb node.