rollup stats_counts metrics correctly
Statsd emits multiple types of metrics:

stats.gauges : gauges
stats.timers : timers
stats.*      : counters expressed as rate-per-second
stats_counts.* : counters expressed as total-per-time-period

Whisper (the storage engine behind graphite) regularly rolls up data
into series with less resolution - eg from 5 second resolution to 10
second resolution.

The function used to perform this rollup is different for the different
types of metrics:

stats.* needs average. If the rate-per-second has been 22 for 5 seconds,
then 44 for 5 seconds, the average rate-per-second over the 10 second
period is 33, the average of 22 and 44. The same applies for gauges and
timers.

stats_counts.* needs sum. If the total events was 110 in the first 5
seconds, then 220 in the next 5 seconds, the total over the 10 second
period was 330, the sum of 110 and 220.

This commit adds the storage-aggregation.conf configuration file to
graphite to set the rollup function correctly.