[#48247755] Don't list passive checks as disabled
Nagios passive checks currently show as "disabled" in the Tactical Overview
page because the active check portion of the service is, indeed, disabled.

Replace the use of `active_checks_enabled 0` with `check_period never` to
prevent them showing as disabled but still never schedule an active check.
This should have the same effect, even when freshness checking isn't used.