[#48247755] Alert if Puppet hasn't run in 2hrs
Set a freshness threshold on Puppet passive checks. Will fire a WARNING if
it doesn't receive any data within 2hrs. This equates to 4 normal runs from
cron.

I'm unilaterally deciding that the value of 24hrs from the original story is
too long. We'd want to know before that. Scheduled downtime should be used
if you expect to fire this.