Per-app nagios CPU usage warnings for Whitehall
The whitehall application is particularly CPU intensive, because
of the number of requests that it services. This commit adds
parameters to the govuk::app class that allows for setting the
thresholds for cpu alerts. The default is the status quo of 150%
for a warning and 200% for a critical.

This commit also raises the whitehall thresholds to 300% and 400%