Revert "[#48247755] Don't list passive checks as disabled"
This reverts commit f26bbbddaff0aa7815d2c05f9679ce2f52e23c9c.

The use of `check_period never` does not work with freshness checks. It
prevents them from ever being scheduled. This used to work in older versions
of Nagios, but not anymore. The freshness checking is more important than
services being displayed as "disabled" in the tactical dashboard.

We could use different settings for passive checks that have a freshness
threshold versus those that don't. But arguably that introduces complexity
that we don't currently need.

Conflicts:
	modules/nagios/templates/passive_service.erb