[#44293669] Feature flag cache real_ip_header
Convert the use of `real_ip_header` on the cache/router nodes to be a
feature flag. This will allow us to set a value provided by our CDN, or any
future upstream L7 load balancer.

The default behaviour, when no value is provided, is to turn off
`real_ip_header`. In which case the `$remote_addr` will (currently) contain
the IP address of our L3 load balancer (vShield Edge). This seems like the
safest behaviour since it's far more dangerous to allow anyone (authed or
not) to spoof their address in our logs.

This raises a problem for Smokey though, which takes advantage of the XFF
spoofing in order to bypass rate limiting. It's also public domain that it
does this - being a public repo on GitHub. I think we need to find a
different way of not subjecting Smokey to rate limiting.